Add cross-platform portability gates and release safety fixes
Introduce a repository-wide portability checker with AppleClang/libc++ syntax coverage, strict GLSL validation, and Windows/MSVC and NTFS compatibility checks. Expose the checks through Make targets and run the compiler-free and toolchain-specific lanes in CI with required-tool enforcement.

Make optimized builds preserve the finiteness checks relied on by gameplay and rendering code, use precise floating-point behavior on MSVC, and enable a focused strict warning set for cross-platform diagnostics. Fix the portability findings in shaders, math constants, evaluation order, range-loop bindings, abstract overrides, ignored results, and enum comparisons, then document the workflow and record the fixes in the changelog.

The branch was formatted with make format and rebased by fast-forwarding from the latest origin/main before this commit.

+++ ui/input_bindings.cpp
@@ -739,7 +739,7 @@ auto InputBindings::format(const Chord& chord) -> QString {
 
   if (chord.mouse_button != 0) {
     for (const auto& entry : k_button_names) {
-      if (chord.mouse_button == entry.button) {
+      if (chord.mouse_button == static_cast<int>(entry.button)) {
         return modifiers_to_string(chord.modifiers) + QString::fromLatin1(entry.name);
       }
     }
@@ -805,7 +805,7 @@ auto InputBindings::describe(const QString& shortcut) -> QString {
   }
   if (chord.mouse_button != 0) {
     for (const auto& entry : k_button_names) {
-      if (chord.mouse_button == entry.button) {
+      if (chord.mouse_button == static_cast<int>(entry.button)) {
         return modifiers_to_string(chord.modifiers) + tr(entry.name);
       }
     }
